chore(SLB-385): refactor responsiveImage resolver
