+++ apps/cms/config/sync/core.entity_form_display.media.image.default.yml
@@ -4,7 +4,7 @@ status: true
 dependencies:
   config:
     - field.field.media.image.field_media_image
-    - image.style.medium
+    - image.style.large
     - media.type.image
   module:
     - focal_point
@@ -26,7 +26,7 @@ content:
     region: content
     settings:
       progress_indicator: throbber
-      preview_image_style: medium
+      preview_image_style: large
       preview_link: false
       offsets: '50,50'
     third_party_settings: {  }
