+++ apps/cms/config/sync/crop.type.focal_point.yml
@@ -4,8 +4,8 @@ status: true
 dependencies: {  }
 _core:
   default_config_hash: flCi9IdafdLXlJqvoHguutUOiC05-aynK4niYN4YZ3o
-id: focal_point
 label: 'Focal point'
+id: focal_point
 description: 'Crop type used by Focal point module.'
 aspect_ratio: ''
 soft_limit_width: null
