+++ packages/schema/package.json
@@ -68,6 +68,8 @@
     "@amazeelabs/scalars": "^1.6.13",
     "@swc/cli": "^0.1.63",
     "@swc/core": "^1.3.102",
+    "gatsby-plugin-sharp": "^5.13.0",
+    "gatsby-source-filesystem": "^5.13.0",
     "image-size": "^1.1.1",
     "mime-types": "^2.1.35"
   }
