+++ packages/schema/src/types/gatsby-plugin-sharp.d.ts
@@ -1,0 +1,3 @@
+declare module 'gatsby-plugin-sharp' {
+  function fluid(props: any): any;
+}
